# license-server/src/main/java/de/netzkronehd/license/model/LicenseCheckResult.java
package de.netzkronehd.license.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.format.annotation.DateTimeFormat;

import java.time.OffsetDateTime;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class LicenseCheckResult {

    private String licenseKey;
    private String publisher;
    private String notes;
    private boolean valid;

    @DateTimeFormat(
            iso = DateTimeFormat.ISO.DATE_TIME
    )
    private OffsetDateTime validUntil;

}